New: More metadata included in manifests.
Wrote some small improvements to add more of the metadata
sourced from archive.org in the manifest response.

A `metadata` list will be generated by iterating through a new
top level `METADATA_FIELDS` tuple and adding any of the found
values to the manifest.

Descriptions and links back to the internet archive are also
included in the manifest that is produced.